+++ resources/ts/browser.ts
@@ -1,0 +1,4 @@
+// @ts-ignore
+const phpRevival = typeof browser === 'undefined' ? chrome : browser
+
+export default phpRevival