+++ changelog.md
@@ -1,5 +1,12 @@
 # Release Notes
 
+## v2.0.3 (2020-02-13)
+
+- Replaced absolute path for images on home page with relative one
+- Added more random videos to a home page sidebar
+
+----
+
 ## v2.0.2 (2020-02-9)
 
 - Removed permissions access from manifest.json